Moving to singleton instance of RPC API
